Fix circular dependency and updated renumbering
