+++ python/cugraph/tests/test_utils.py
@@ -85,8 +85,11 @@ def test_get_traversed_cost(graph_file):
     # run SSSP starting at vertex 17
     df = cugraph.sssp(G,  16)
 
-    answer = cugraph.utils.get_traversed_cost(df, 16, cu_M['0'], cu_M['1'],
-                                              cu_M['info'])
+    answer = cugraph.utilities.path_retrieval.get_traversed_cost(df, 16,
+                                                                   cu_M['0'],
+                                                                   cu_M['1'],
+                                                                   cu_M['info']
+                                                                   )
 
     df = df.sort_values(by='vertex').reset_index()
     answer = answer.sort_values(by='vertex').reset_index()
